refactor: Slidevテーマの背景コンポーネントを分離してコード構造を改善
- global-bottom.vueを945行から97行に削減
- OceanBackground.vue: Three.js海洋アニメーション (459行)
- NeonBackground.vue: ネオンポリゴンアニメーション (234行)
- ThemeIndicator.vue: テーマ切り替えUI (151行)
- useNeonThemes.ts: Neonテーマ設定をcomposableに分離
- 2025-08-07-zenoプロジェクトの依存関係を修正

各コンポーネントが単一責任の原則に従い、保守性とテスト可能性が向上
