+++ apps/slides/2025-08-07-zeno/package.json
@@ -13,11 +13,12 @@
     "@slidev/cli": "^0.50.0-beta.6",
     "@slidev/theme-default": "^0.25.0",
     "@slidev/theme-seriph": "^0.25.0",
-    "vue": "^3.5.12",
-    "slidev-theme-mozumasu": "file:../../../packages/ui/slidev-theme-mozumasu"
+    "@slidev/types": "^0.50.0-beta.6",
+    "slidev-theme-mozumasu": "file:../../../packages/ui/slidev-theme-mozumasu",
+    "vue": "^3.5.12"
   },
   "dependencies": {
-    "three": "^0.179.1",
-    "@mozumasu/shared-components": "workspace:*"
+    "@mozumasu/shared-components": "workspace:*",
+    "three": "^0.179.1"
   }
 }
